requirements.txt
Django

django-mssql-backend
pyodbc
mssql-django
djangorestframework
python-dotenv
requests